Use putShort() for inserting TAG_MODE into bundle
Since TAG_MODE is a short value,
putShort() should be used to insert a value for it
instead of putInt().

This fixes the following warning:

"[...] com.manichord.mgit W/Bundle: Key mode expected Short but value was
a java.lang.Integer.  The default value 0 was returned."

+++ src/main/java/me/sheimi/sgit/activities/ViewFileActivity.java
@@ -88,7 +88,7 @@ public class ViewFileActivity extends SheimiFragmentActivity {
         mFileFragment = new ViewFileFragment();
         mFileFragment.setArguments(b);
         mActivityMode = extras.getShort(TAG_MODE, TAG_MODE_NORMAL);
-        b.putInt(TAG_MODE, mActivityMode);
+        b.putShort(TAG_MODE, mActivityMode);
         getActionBar().setDisplayHomeAsUpEnabled(true);
         setTitle(new File(fileName).getName());
     }
